+++ aqueduct/test/http/compilation_errors/resource_controller_ambiguous_operation_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import "dart:core";
 
 import 'package:aqueduct/aqueduct.dart';
-import 'package:aqueduct/src/runtime/mirror_impl.dart';
+import 'package:aqueduct/src/runtime/runtime_impl.dart';
 import "package:test/test.dart";
 
 void main() {
