+++ aqueduct/test/http/compilation_errors/resource_controller_bind_dynamic_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:aqueduct/aqueduct.dart';
-import 'package:aqueduct/src/runtime/mirror_impl.dart';
+import 'package:aqueduct/src/runtime/runtime_impl.dart';
 import "package:test/test.dart";
 
 void main() {
